requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.4
django-allauth==0.43.0
django-debug-toolbar==3.2
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
Markdown==3.3.4
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
